+++ include/tad/transporte.hpp
@@ -9,15 +9,15 @@ private:
     // Dados que identificam este transporte específico
     int m_id_origem;
     int m_id_destino;
-    double m_tempo_partida;
-    double m_tempo_chegada_previsto;
+    int m_tempo_partida;
+    int m_tempo_chegada_previsto;
 
     // Lista de ponteiros para todos os pacotes neste "lote" de transporte
     ListaDinamica<Pacote*> m_pacotes_neste_transporte;
 
 public:
     // Construtor para inicializar as informações do transporte
-    Transporte(int origem, int destino, double partida, double chegada);
+    Transporte(int origem, int destino, int partida, int chegada);
 
     // Adiciona um ponteiro de pacote a este transporte
     void adicionarPacote(Pacote* pacote);
@@ -25,8 +25,8 @@ public:
     // Getters para acessar as informações do transporte
     int getOrigemID() const;
     int getDestinoID() const;
-    double getTempoPartida() const;
-    double getTempoChegadaPrevisto() const;
+    int getTempoPartida() const;
+    int getTempoChegadaPrevisto() const;
 
     // Retorna uma referência constante à lista de pacotes
     const ListaDinamica<Pacote*>& getPacotes() const;
